requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
supabase
google-generativeai
requests
python-dotenv
